refactor: combine storage and state class
